fixes comparison error
clang 10 complained about comparing int to unsigned int

+++ include/progresscpp/ProgressBar.hpp
@@ -31,7 +31,7 @@ public:
 
         std::cout << "[";
 
-        for (int i = 0; i < bar_width; ++i) {
+        for (long i = 0; i < long(bar_width); ++i) {
             if (i < pos) std::cout << complete_char;
             else if (i == pos) std::cout << ">";
             else std::cout << incomplete_char;
